apps/cvs/cv_service/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=1.10.7
sqlalchemy>=2.0.9
psycopg2-binary>=2.9.6
alembic>=1.10.3
python-jose>=3.3.0
python-multipart>=0.0.6
passlib>=1.7.4
bcrypt>=4.0.1
pyjwt>=2.6.0
python-dotenv>=1.0.0
asyncpg>=0.27.0
pillow>=10.0.0
python-magic>=0.4.27
pydantic-extra-types==2.0.0
email-validator==2.0.0.post2
pyyaml==6.0.1
jsonschema==4.18.6
python-magic-bin; sys_platform == 'win32'
python-magic; sys_platform != 'win32'
pypdf
python-docx 